+++ frontend/app/upload/components/Crawler/hooks/useCrawler.ts
@@ -3,8 +3,8 @@ import { redirect } from "next/navigation";
 import { useCallback, useRef, useState } from "react";
 
 import { useSupabase } from "@/lib/context/SupabaseProvider";
+import { useAxios } from "@/lib/hooks";
 import { useToast } from "@/lib/hooks/useToast";
-import { useAxios } from "@/lib/useAxios";
 
 import { isValidUrl } from "../helpers/isValidUrl";
